refactor(widget-plugin-filter, widget-kit-web): move return-meta file
