+++ packages/shared/widget-kit-web/src/util.ts
@@ -3,6 +3,5 @@ export * from "./util/ensure.js";
 export * from "./util/generate-uuid.js";
 export * from "./util/get-dimensions.js";
 export * from "./util/parse-style.js";
-export * from "./util/result-meta.js";
 export * from "./util/throttle.js";
 export * from "./util/tuple.js";
