+++ packages/shared/widget-plugin-filtering/src/result-meta.ts
